chore(gpmf): review cleanup
Various cleanups spotted in review.
