+++ pkg/gopro/gpmf/interfaces.go
@@ -8,3 +8,13 @@ import (
 type offseter interface {
 	offsets(start, end time.Duration)
 }
+
+// number represents a number we can scale.
+type number interface {
+	int8 | uint8 |
+		int16 | uint16 |
+		int32 | uint32 |
+		float64 | float32 |
+		int64 | uint64 |
+		Int16_16 | Int32_32
+}
